JZL1CSS-40 - Rearrange project classes and packages
