+++ src/main/java/com/infoshare/academy/Main.java
@@ -1,10 +1,10 @@
 package com.infoshare.academy;
 
-import com.infoshare.academy.menu.Menu;
+import com.infoshare.academy.reservation.ReservationControl;
 
 public class Main {
 
     public static void main(String[] args) {
-        Menu.getUserDataInput();
+        ReservationControl.getUserDataInput();
     }
 }
